Move backend chat live tests
